TASK: speed up Docker build by caching Golang
As described in [1], we can use cache mounts to
speed up the Golang build. This means we do not
download all go dependencies all the time, and
also keep intermediate compilation artifacts.

This has no downsides IMHO, because the Go compiler
automatically detects if the artifacts are stale or not.

[1] https://docs.docker.com/build/cache/optimize/#use-cache-mounts